connect app with rockit main,music vol, mics for treble,bass,mute
1. change the volume to float with real range -100-0db.
2. change the bass, treble real range to -25-25db
2. rockit update to v1.05

+++ app/pbox_app.h
@@ -54,8 +54,8 @@ typedef struct {
         placement_t placement;
         stereo_mode_t stereo;
         inout_door_t outdoor;
-        uint32_t mainVolumeLevel;//main volume
-        uint32_t musicVolumeLevel;//music volume
+        float mainVolumeLevel;//main volume
+        float musicVolumeLevel;//music volume
         mic_data_t micData[MIC_NUM];//mMicVolumeLevel
         uint32_t accomLevel;
         uint32_t humanLevel;
@@ -95,7 +95,7 @@ void pbox_app_music_trackid(uint32_t id, display_t policy);
 void pbox_app_music_start(display_t policy);
 void pbox_app_music_resume(display_t policy);
 void pbox_app_music_stop(display_t policy);
-void pbox_app_music_set_volume(uint32_t volume, display_t policy);
+void pbox_app_music_set_volume(float volume, display_t policy);
 void pbox_app_music_volume_up(display_t policy);
 void pbox_app_music_volume_down(display_t policy);
 void pbox_app_music_album_next(bool next, display_t policy);
@@ -104,7 +104,7 @@ void pbox_app_music_original_singer_open(bool original, display_t policy);
 void pbox_app_music_album_loop(uint32_t mode, display_t policy);
 void pbox_app_music_seek_position(uint32_t dest, uint32_t duration, display_t policy);
 char* pbox_app_usb_get_title(uint32_t trackId);
-void pbox_app_music_set_mic_volume(uint32_t micNum, uint32_t volume, display_t policy);
+void pbox_app_music_set_mic_volume(uint32_t micNum, float volume, display_t policy);
 void pbox_app_music_set_mic_mute(bool mute, display_t policy);
 void pbox_app_music_mics_init(display_t policy);
 void pbox_app_music_set_accomp_music_level(uint32_t volume, display_t policy);
@@ -113,9 +113,9 @@ void pbox_app_music_set_human_music_level(uint32_t volume, display_t policy);
 void pbox_app_music_set_reserv_music_level(uint32_t volume, display_t policy);
 void pbox_app_music_set_echo_3a(bool enable, display_t policy);
 void pbox_app_music_set_recoder_revert(pbox_revertb_t reverbMode, display_t policy);
-void pbox_app_music_set_mic_treble(uint8_t index, uint32_t treble, display_t policy);
-void pbox_app_music_set_mic_bass(uint8_t index, uint32_t bass, display_t policy);
-void pbox_app_music_set_mic_reverb(uint8_t index, uint32_t reverb, display_t policy);
+void pbox_app_music_set_mic_treble(uint8_t index, float treble, display_t policy);
+void pbox_app_music_set_mic_bass(uint8_t index, float bass, display_t policy);
+void pbox_app_music_set_mic_reverb(uint8_t index, float reverb, display_t policy);
 void pbox_version_print(void);
 
 void pbox_app_uac_state_change(uac_role_t role, bool start, display_t policy);
@@ -145,7 +145,7 @@ void pbox_app_btsoc_set_human_voice_fadeout(bool fadeout, display_t policy);
 void pbox_app_btsoc_get_input_source(display_t policy);
 void pbox_app_music_set_input_source(input_source_t source, play_status_t status, display_t policy);
 void pbox_app_music_get_music_volume(display_t policy);
-void pbox_app_music_set_music_volume(uint32_t volume, display_t policy);
+void pbox_app_music_set_music_volume(float volume, display_t policy);
 
 bool is_dest_source_switchable(input_source_t source, switch_source_t mode);
 bool is_input_source_selected(input_source_t source, switch_source_t mode);
