+++ app/pbox_common.h
@@ -193,10 +193,10 @@ typedef enum {
 typedef struct {
     uint8_t index;
     mic_mux_t micMux;
-    uint32_t micVolume;
-    uint32_t micTreble;
-    uint32_t micBass;
-    uint32_t micReverb;
+    float micVolume;
+    float micTreble;
+    float micBass;
+    float micReverb;
 } mic_data_t;
 
 typedef enum {
